feat: :sparkles: Upload genome to firebase

+++ src/server.ts
@@ -1,25 +1,35 @@
 import { Worker } from 'worker_threads';
 import express from "express";
+import { uploadNewGenome } from './firebaseControl.js';
+
 const app = express();
 
 const PORT = 3000;
 const TIMEOUT = 1.5 * 60 * 1000;
 
 async function runWorker(workerData: any) {
+	if (workerData.length !== 10) {
+		return "No parameters provided!";
+	}
+
 	return new Promise((resolve, reject) => {
 		const worker = new Worker("./dist/index.js", {
 			workerData: workerData
 		});
 
-		const timeout = setTimeout(() => {
+		const timeout = setTimeout(async () => {
 			console.log("process timed out :/");
 			worker.terminate();
-			resolve(`Something went wrong :/`)
+			resolve(await runWorker(workerData));
 		}, TIMEOUT)
 
 		worker.on("online", () => console.log("New worker started!"));
 		worker.on("message", (v) => {
-			if (v?.[0]) resolve(v[1]);
+			if (v?.[0]) {
+				clearTimeout(timeout);
+				uploadNewGenome(v[1]);
+				resolve(v[1]);
+			}
 		});
 		worker.on("exit", () => console.log("Worker stopped!"))
 	})
